added interface class for app refactored web and webApi -> now RestApi.h fix calcSunrise fixed calcSunrise trigger calculation display zero values on /live added changes from #483
