+++ src/utils/helper.h
@@ -18,6 +18,7 @@ namespace ah {
     void ip2Char(uint8_t ip[], char *str);
     double round3(double value);
     String getDateTimeStr(time_t t);
+    uint64_t Serial2u64(const char *val);
 }
 
 #endif /*__HELPER_H__*/
